change to size_t at counter of clockManager.

+++ src/clockManager.hpp
@@ -7,7 +7,7 @@ class clockManager {
     int64_t log[5];
     size_t logLen;
     volatile size_t write_index = 0;
-    volatile int counter = 0;
+    volatile size_t counter = 0;
     uint16_t div;
     uint32_t div_base;
     uint16_t HZ;
